attributes editor

+++ src/dbmpPlus/DBMPPlus.java
@@ -1,46 +1,36 @@
 package dbmpPlus;
 
 import java.io.File;
-import java.io.InputStream;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.NoSuchElementException;
-
 import javafx.application.Application;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
-import javafx.scene.image.Image;
 import javafx.scene.input.KeyCode;
 import javafx.scene.layout.*;
-import javafx.scene.text.Text;
 import javafx.stage.FileChooser;
 import javafx.stage.Stage;
-import javafx.util.Callback;
 
 public class DBMPPlus extends Application {
 	
 	public static DBMP mainDBMP = new DBMP();
 	
-//	public static VBox partsDisplay;
-	public static ListView<Part>  partsDisplay;
+	public static ListView<Part> partsDisplay;
+//	public static ListView<Attribute> attributesDisplay;
 	public static VBox attributesDisplay;
 	
 	public static ArrayList<CarPartListCell> carPartListCells = new ArrayList<CarPartListCell>();
 
 	public static boolean debug = true;
 	
-	
 	public static void main(String[] args) {
-
 		//if wrong that will throw an exception due to trying to spawn an error window without having initialized javafx 
 		mainDBMP = DBMP.loadDBMP(new File("C:\\Users\\gaupp\\OneDrive\\Documents\\z NFS MODDING\\z bordel\\LOT_ELI_111_06.bin"));
 //		mainDBMP = DBMP.loadDBMP(new File("C:\\Users\\NI240SX\\Documents\\NFS\\a MUCP\\voitures\\z done\\car bmw e92\\dbmp step 8.bin"));
         launch(args);
 	}
 	
-
     public void start(Stage primaryStage) {
         primaryStage.setTitle("fire");
 
@@ -156,91 +146,79 @@ public class DBMPPlus extends Application {
         	});
         	
         }
-
-        
         
         windowTop.getChildren().addAll(menuBar, shortcutsBar);
         
-        // Car parts zone
-//        partsDisplay = new VBox();
-//        ScrollPane scrollPaneParts = new ScrollPane(partsDisplay);
-//        scrollPaneParts.setFitToWidth(true);
-//        scrollPaneParts.setFitToHeight(true);
-
         partsDisplay = new ListView<Part>();
-        partsDisplay.setCellFactory( lv -> {//new Callback<ListView<Part>, ListCell<Part>>() {
+        partsDisplay.setCellFactory( lv -> {
         	 CarPartListCell cell = new CarPartListCell() {
                  @Override
                  protected void updateItem(Part item, boolean empty) {
                      super.updateItem(item, empty);
-//                     setText(item);
                  }
              };
              return cell;
-//            @Override
-//            public CarPartListCell call(ListView<Part> listView) {
-//                return new CarPartListCell();
-//            }
         });
-        
         partsDisplay.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
-        
         partsDisplay.setOnKeyPressed(e -> {
-        	if (e.getCode().equals(KeyCode.DELETE)) {
-        		deleteSelectedParts();
-        	}
+        	if (e.getCode().equals(KeyCode.DELETE)) deleteSelectedParts();
         	if (e.getCode().equals(KeyCode.Z) && e.isControlDown()) Undo.undo();
         	for (CarPartListCell c : DBMPPlus.carPartListCells) if(c!=null && c.getItem()!=null) c.checkBox.setSelected(DBMPPlus.partsDisplay.getSelectionModel().getSelectedItems().contains(c.getItem()));
+        	
+        	DBMPPlus.attributesDisplay.getChildren().clear();
+        	for (Attribute a: partsDisplay.getSelectionModel().getSelectedItem().attributes) DBMPPlus.attributesDisplay.getChildren().add(a.dataHBox);
+        	
         	e.consume();
         });
         
-        
+//        attributesDisplay = new ListView<Attribute>();
+//        attributesDisplay.setCellFactory( lv -> {
+//        	  AttributeListCell cell = new AttributeListCell() {
+//                 @Override
+//                 protected void updateItem(Attribute item, boolean empty) {
+//                     super.updateItem(item, empty);
+//                 }
+//             };
+//             return cell;
+//        });
+//        attributesDisplay.setOnKeyPressed(e -> {
+//        	if (e.getCode().equals(KeyCode.Z) && e.isControlDown()) Undo.undo();
+//        	e.consume();
+//        });
+//        
         attributesDisplay = new VBox();
         ScrollPane scrollPaneAttrib = new ScrollPane(attributesDisplay);
         scrollPaneAttrib.setFitToWidth(true);
-        scrollPaneAttrib.setFitToHeight(true);
-
+        
+        
         for (Part p : mainDBMP.parts) {
-//            partsDisplay.getChildren().add(p);
         	partsDisplay.getItems().add(p);
-        	
         }
 
-        // Status Bar
-        Label statusBar = new Label("Status: Ready");
+//        Label statusBar = new Label("Status: Ready");
 
-        // Layout
         BorderPane root = new BorderPane();
         root.setTop(windowTop);
-//        root.setCenter(scrollPaneParts);
         root.setCenter(partsDisplay);
+//        root.setRight(attributesDisplay);
         root.setRight(scrollPaneAttrib);
-        root.setBottom(statusBar);
+//        root.setBottom(statusBar);
+        
 
+//        attributesDisplay.setMinWidth(root.getWidth()/2);
+        
+        
         Scene scene = new Scene(root, 1024, 600);
         primaryStage.setScene(scene);
         primaryStage.show();
     }
 
-    // Create a custom list item with a checkbox, name, and editable text field
-/*    private HBox createListItem(String itemName) {
-        CheckBox checkBox = new CheckBox();
-        TextField textField = new TextField(itemName);
-
-        HBox listItem = new HBox(10);
-        listItem.getChildren().addAll(checkBox, textField);
-
-        return listItem;
-    }*/
-	
-    public static void updateAllPartsDisplay() {
+    public static void updateAllPartsDisplay() { //broken for some weird reason
+    	mainDBMP.updateAll();
     	partsDisplay.getItems().clear();
     	for (Part p : mainDBMP.parts) {
        	partsDisplay.getItems().add(p);
-       	
-       }
-    	
-    	
+    	}
     }
     
     public static void deleteSelectedParts() {
@@ -253,11 +231,5 @@ public class DBMPPlus extends Application {
     		DBMPPlus.mainDBMP.parts.remove(toDelete[i]);
 		}
 		for (CarPartListCell c : DBMPPlus.carPartListCells) if(c!=null && c.getItem()!=null) c.checkBox.setSelected(DBMPPlus.partsDisplay.getSelectionModel().getSelectedItems().contains(c.getItem()));
-
 	}
-/*    	partsDisplay.getChildren().clear();
-    	for (Part p : mainDBMP.parts) {
-    		partsDisplay.getChildren().add(p);
-    	}
-    }*/
 }
